requirements.txt
streamlit
tensorflow
librosa
soundfile
numpy
google-generativeai
